Add files via upload
Linked List is a part of the Collection framework present in java.util package. This class is an implementation of the LinkedList data structure, which is a linear data structure where the elements are not stored in contiguous locations, and every element is a separate object with a data part and an address part. The elements are linked using pointers and addresses, and each element is known as a node. 

Note: The nodes cannot be accessed directly instead we have to start from the head and follow the link until we find the node that we want.

Normal List vs Doubly LinkedList:

The main difference between a normal linked list and a doubly LinkedList is that Doubly linked list contains an extra pointer which is known as previous pointer. In this each node points to both the next and previous nodes.